Hotbar完成

+++ src/main/java/net/ccbluex/liquidbounce/ui/client/hud/element/elements/Hotbar.java
@@ -53,7 +53,7 @@ public class Hotbar extends Element {
             GlStateManager.pushMatrix();
             GlStateManager.scale(scale, scale, scale);
             float x = i * 25f / scale - scale * 2f;
-            now.renderHotbarItem(i, (int) x + scale, (int) (now.translate.getY() - scale / 2f - 2f / scale), mc.timer.renderPartialTicks);
+            now.renderHotbarItem(i, x + scale,now.translate.getY() - scale / 2f - 2f / scale, mc.timer.renderPartialTicks);
             GlStateManager.popMatrix();
 
             mc.getRenderItem().renderItemOverlays(scale == 1f ? Fonts.font15 : Fonts.font18, mc.thePlayer.inventory.mainInventory[i], (int) (i * 25f), -3);
